New script (model_evaluation_SARS2_multicoupling.py) to couple all the models described in a YAML The parameters that the YAML file must contain are:     {'Model_Name': {       'class_weight': 'balanced',       'des': 'Mordred', # indicate which descriptor is used, if it's not used then the parameter is None       'fp': None, # indicate which fingerprint is used, if it's not used then the parameter is None       'percentile_des_ls': [10, 20],       'percentile_fp_ls': [0, 0],       'type': 'svm',       'parameters': XX} # parameters need to run the `type` technique, ie, lr, optimizer... for NN or gamma, C... for SVM
the elements in percentile_fp_ls and percentile_des_ls are used to select the percentage of each feature to be kept, for each percentile explotation.
 ie. in this case we will run first 10% fp + 0% des, and second 20% fp + 0% des.

Some functions were added to small_datasets_utils_coupled.py in order to run the multicoupling.
Also, show_plot parameter in functions plot_results_split and plot_results_CV has been removed, so older versions of this script may have problems with this parameters.